moved code to plugin form
Also fixed blueprints related to code
