+++ .idea/.idea.RPG/.idea/workspace.xml
@@ -18,7 +18,6 @@
         <option value="$PROJECT_DIR$/Source/RPG/Enemy/EnemyBase.h" />
         <option value="$PROJECT_DIR$/Source/RPG/Enemy/EnemyBase.cpp" />
         <option value="$PROJECT_DIR$/Source/RPG/InteractionInterface.h" />
-        <option value="$PROJECT_DIR$/Source/RPG/Quest/Quest.h" />
         <option value="$PROJECT_DIR$/Source/RPG/Weapon/WeaponInfo.h" />
         <option value="$PROJECT_DIR$/Source/RPG/Items/ItemInfo.h" />
         <option value="$PROJECT_DIR$/Source/RPG/RPGCharacter.h" />
@@ -26,8 +25,22 @@
         <option value="$PROJECT_DIR$/Source/RPG/Abilities/AbilityInfo.h" />
         <option value="$PROJECT_DIR$/Source/RPG/Items/ItemSpecialEffect.cpp" />
         <option value="$PROJECT_DIR$/Source/RPG/Items/ItemSpecialEffect.h" />
-        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/RPGCharacterBase.h" />
         <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/RPGCore.Build.cs" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Private/EnemyCharacterBase.cpp" />
+        <option value="$PROJECT_DIR$/Source/RPG/Quest/Quest.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/Quest/QuestInfo.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Private/Quest/QuestInfo.cpp" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Private/Item/ItemInfo.cpp" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/Item/SpecialEffect.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Private/Item/SpecialEffect.cpp" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/Ability/AbilityInfo.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Private/Ability/AbilityInfo.cpp" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/Interaction.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/EnemyCharacterBase.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/Item/ItemInfo.h" />
+        <option value="$PROJECT_DIR$/Source/RPG/RPGGameMode.cpp" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Public/RPGCharacterBase.h" />
+        <option value="$PROJECT_DIR$/Plugins/RPGCore/Source/RPGCore/Private/RPGCharacterBase.cpp" />
       </list>
     </option>
   </component>
