--- Config/DefaultEditor.ini
@@ -1,16 +1,0 @@
-[UnrealEd.SimpleMap]
-SimpleMapName=/Game/2DSideScrollerCPP/Maps/2DSideScrollerExampleMap
-
-[EditoronlyBP]
-bAllowClassAndBlueprintPinMatching=true
-bReplaceBlueprintWithClass= true
-bDontLoadBlueprintOutsideEditor= true
-bBlueprintIsNotBlueprintType= true
-
-[/Script/UnrealEd.LevelEditor2DSettings]
-bEnable2DWidget=True
-bEnableSnapLayers=True
-SnapAxis=Y
-+SnapLayers=(Name="Foreground",Depth=100.000000)
-+SnapLayers=(Name="Default",Depth=0.000000)
-+SnapLayers=(Name="Background",Depth=-400.000000)
