--- Config/DefaultEditorPerProjectUserSettings.ini
@@ -1,2 +1,0 @@
-[ContentBrowser]
-ContentBrowserTab1.SelectedPaths=/Game/2dSideScroller
