+++ Config/DefaultEngine.ini
@@ -1,139 +1,5 @@
 
 
 [/Script/EngineSettings.GameMapsSettings]
-EditorStartupMap=/Game/Maps/test.test
-LocalMapOptions=
-TransitionMap=
-bUseSplitscreen=True
-TwoPlayerSplitscreenLayout=Horizontal
-ThreePlayerSplitscreenLayout=FavorTop
-GameInstanceClass=/Script/Engine.GameInstance
-GameDefaultMap=/Game/Maps/test.test
-ServerDefaultMap=/Game/2DSideScrollerCPP/Maps/2DSideScrollerExampleMap
 GlobalDefaultGameMode=/Script/RPG.RPGGameMode
-GlobalDefaultServerGameMode=None
-
-[/Script/IOSRuntimeSettings.IOSRuntimeSettings]
-MinimumiOSVersion=IOS_11
-
-[/Script/HardwareTargeting.HardwareTargetingSettings]
-TargetedHardwareClass=Desktop
-AppliedTargetedHardwareClass=Desktop
-DefaultGraphicsPerformance=Scalable
-AppliedDefaultGraphicsPerformance=Scalable
-
-[/Script/Engine.Engine]
-+ActiveGameNameRedirects=(OldGameName="TP_2DSideScroller",NewGameName="/Script/RPG")
-+ActiveGameNameRedirects=(OldGameName="/Script/TP_2DSideScroller",NewGameName="/Script/RPG")
-+ActiveGameNameRedirects=(OldGameName="/Script/TP_ThirdPerson",NewGameName="/Script/RPG")
-+ActiveGameNameRedirects=(OldGameName="TP_ThirdPerson",NewGameName="/Script/RPG")
-+ActiveClassRedirects=(OldClassName="TP_2DSideScrollerGameMode",NewClassName="RPGGameMode")
-+ActiveClassRedirects=(OldClassName="TP_2DSideScrollerCharacter",NewClassName="RPGCharacter")
-
-[/Script/Engine.RendererSettings]
-r.Mobile.DisableVertexFog=True
-r.Shadow.CSM.MaxMobileCascades=2
-r.MobileMSAA=1
-r.Mobile.UseLegacyShadingModel=False
-r.Mobile.AllowDitheredLODTransition=False
-r.Mobile.AllowSoftwareOcclusion=False
-r.Mobile.VirtualTextures=False
-r.DiscardUnusedQuality=False
-r.AllowOcclusionQueries=True
-r.MinScreenRadiusForLights=0.030000
-r.MinScreenRadiusForDepthPrepass=0.030000
-r.MinScreenRadiusForCSMDepth=0.010000
-r.PrecomputedVisibilityWarning=False
-r.TextureStreaming=True
-Compat.UseDXT5NormalMaps=False
-r.VirtualTextures=False
-r.VirtualTexturedLightmaps=False
-r.VT.TileSize=128
-r.VT.TileBorderSize=4
-r.vt.FeedbackFactor=16
-r.VT.EnableCompressZlib=True
-r.VT.EnableCompressCrunch=False
-r.ClearCoatNormal=False
-r.AnisotropicBRDF=False
-r.ReflectionCaptureResolution=128
-r.ReflectionEnvironmentLightmapMixBasedOnRoughness=True
-r.ForwardShading=False
-r.VertexFoggingForOpaque=True
-r.AllowStaticLighting=True
-r.NormalMapsForStaticLighting=False
-r.GenerateMeshDistanceFields=False
-r.DistanceFieldBuild.EightBit=False
-r.GenerateLandscapeGIData=False
-r.DistanceFieldBuild.Compress=False
-r.TessellationAdaptivePixelsPerTriangle=48.000000
-r.SeparateTranslucency=True
-r.TranslucentSortPolicy=0
-TranslucentSortAxis=(X=0.000000,Y=-1.000000,Z=0.000000)
-r.CustomDepth=1
-r.CustomDepthTemporalAAJitter=True
-r.PostProcessing.PropagateAlpha=0
-r.DefaultFeature.Bloom=True
-r.DefaultFeature.AmbientOcclusion=True
-r.DefaultFeature.AmbientOcclusionStaticFraction=True
-r.DefaultFeature.AutoExposure=False
-r.DefaultFeature.AutoExposure.Method=0
-r.DefaultFeature.AutoExposure.Bias=1.000000
-r.DefaultFeature.AutoExposure.ExtendDefaultLuminanceRange=False
-r.UsePreExposure=True
-r.EyeAdaptation.EditorOnly=False
-r.DefaultFeature.MotionBlur=False
-r.DefaultFeature.LensFlare=False
-r.TemporalAA.Upsampling=False
-r.SSGI.Enable=False
-r.DefaultFeature.AntiAliasing=0
-r.DefaultFeature.LightUnits=1
-r.DefaultBackBufferPixelFormat=4
-r.Shadow.UnbuiltPreviewInGame=True
-r.StencilForLODDither=False
-r.EarlyZPass=3
-r.EarlyZPassOnlyMaterialMasking=False
-r.DBuffer=True
-r.ClearSceneMethod=1
-r.BasePassOutputsVelocity=False
-r.VertexDeformationOutputsVelocity=False
-r.SelectiveBasePassOutputs=False
-bDefaultParticleCutouts=False
-fx.GPUSimulationTextureSizeX=1024
-fx.GPUSimulationTextureSizeY=1024
-r.AllowGlobalClipPlane=False
-r.GBufferFormat=1
-r.MorphTarget.Mode=True
-r.GPUCrashDebugging=False
-vr.InstancedStereo=False
-r.MobileHDR=True
-vr.MobileMultiView=False
-r.Mobile.UseHWsRGBEncoding=False
-vr.RoundRobinOcclusion=False
-vr.ODSCapture=False
-r.MeshStreaming=False
-r.WireframeCullThreshold=5.000000
-r.RayTracing=False
-r.RayTracing.UseTextureLod=False
-r.SupportStationarySkylight=True
-r.SupportLowQualityLightmaps=True
-r.SupportPointLightWholeSceneShadows=True
-r.SupportAtmosphericFog=True
-r.SupportSkyAtmosphere=True
-r.SupportSkyAtmosphereAffectsHeightFog=False
-r.SkinCache.CompileShaders=False
-r.SkinCache.DefaultBehavior=1
-r.SkinCache.SceneMemoryLimitInMB=128.000000
-r.Mobile.EnableStaticAndCSMShadowReceivers=True
-r.Mobile.EnableMovableLightCSMShaderCulling=True
-r.Mobile.AllowDistanceFieldShadows=True
-r.Mobile.AllowMovableDirectionalLights=True
-r.MobileNumDynamicPointLights=4
-r.MobileDynamicPointLightsUseStaticBranch=True
-r.Mobile.EnableMovableSpotlights=False
-r.GPUSkin.Support16BitBoneIndex=False
-r.GPUSkin.Limit2BoneInfluences=False
-r.SupportDepthOnlyIndexBuffers=True
-r.SupportReversedIndexBuffers=True
-r.SupportMaterialLayers=False
-r.LightPropagationVolume=False
