--- Config/DefaultGame.ini
@@ -1,8 +1,0 @@
-[/Script/EngineSettings.GeneralProjectSettings]
-ProjectID=ED318B264BC52045991B1697087510A1
-ProjectName=2D Side Scroller Game Template
-
-[/Script/UnrealEd.ProjectPackagingSettings]
-bCompressed=True
-bSkipEditorContent=True
-
