+++ Plugins/RPGCore/Source/RPGCore/Private/Interaction.cpp
@@ -1,0 +1,6 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "Interaction.h"
+
+// Add default functionality here for any IInteraction functions that are not pure virtual.
