+++ Plugins/RPGCore/Source/RPGCore/Private/Item/ItemInfo.cpp
@@ -1,5 +1,5 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "ItemInfo.h"
+#include "RPGCore/Public/Item/ItemInfo.h"
