+++ Plugins/RPGCore/Source/RPGCore/Private/Item/SpecialEffect.cpp
@@ -1,0 +1,12 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "RPGCore/Public/Item/SpecialEffect.h"
+
+void ASpecialEffect::ApplyEffect_Implementation(AActor* CreatorOfEffect, FVector Origin)
+{
+}
+
+void ASpecialEffect::OnEffectEnds_Implementation()
+{
+}
