--- Plugins/RPGCore/Source/RPGCore/Private/Quest/AbilityInfo.cpp
@@ -1,5 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "AbilityInfo.h"
-
