--- Plugins/RPGCore/Source/RPGCore/Public/Quest/AbilityInfo.h
@@ -1,17 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Engine/UserDefinedStruct.h"
-#include "AbilityInfo.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class RPGCORE_API UAbilityInfo : public UUserDefinedStruct
-{
-	GENERATED_BODY()
-	
-};
