--- Source/RPG/Enemy/EnemyBase.h
@@ -1,18 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "PaperCharacter.h"
-#include "EnemyBase.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class RPG_API AEnemyBase : public APaperCharacter
-{
-	GENERATED_BODY()
-	public:
-	AEnemyBase();
-};
