--- Source/RPG/InteractionInterface.cpp
@@ -1,6 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "InteractionInterface.h"
-
-// Add default functionality here for any IInteractionInterface functions that are not pure virtual.
