--- Source/RPG/Items/ItemSpecialEffect.cpp
@@ -1,12 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "ItemSpecialEffect.h"
-
-void AItemSpecialEffect::ApplyEffect_Implementation(AActor* CreatorOfEffect, FVector Origin)
-{
-}
-
-void AItemSpecialEffect::OnEffectEnds_Implementation()
-{
-}
