--- Source/RPG/Levels/LevelScriptActorBase.h
@@ -1,17 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Engine/LevelScriptActor.h"
-#include "LevelScriptActorBase.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class RPG_API ALevelScriptActorBase : public ALevelScriptActor
-{
-	GENERATED_BODY()
-	
-};
