+++ Source/RPG/RPGGameMode.cpp
@@ -1,10 +1,9 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
 #include "RPGGameMode.h"
-#include "RPGCharacter.h"
 
 ARPGGameMode::ARPGGameMode()
 {
 	// Set default pawn class to our character
-	DefaultPawnClass = ARPGCharacter::StaticClass();	
+	
 }
